+++ genium/src/test/resources/smoke/dates/output/cbridge/src/smoke/cbridge_Dates.cpp
@@ -18,7 +18,7 @@ _baseRef smoke_Dates_copy_handle(_baseRef handle) {
         : 0;
 }
 _baseRef
-smoke_Dates_DateStruct_create_handle( int64_t dateField )
+smoke_Dates_DateStruct_create_handle( double dateField )
 {
     ::smoke::Dates::DateStruct* _struct = new ( std::nothrow ) ::smoke::Dates::DateStruct();
     _struct->date_field = Conversion<std::chrono::system_clock::time_point>::toCpp( dateField );
@@ -30,7 +30,7 @@ smoke_Dates_DateStruct_release_handle( _baseRef handle )
     delete get_pointer<::smoke::Dates::DateStruct>( handle );
 }
 _baseRef
-smoke_Dates_DateStruct_create_optional_handle(int64_t dateField)
+smoke_Dates_DateStruct_create_optional_handle(double dateField)
 {
     auto _struct = new ( std::nothrow ) ::genium::optional<::smoke::Dates::DateStruct>( ::smoke::Dates::DateStruct( ) );
     (*_struct)->date_field = Conversion<std::chrono::system_clock::time_point>::toCpp( dateField );
@@ -44,19 +44,19 @@ smoke_Dates_DateStruct_unwrap_optional_handle( _baseRef handle )
 void smoke_Dates_DateStruct_release_optional_handle(_baseRef handle) {
     delete reinterpret_cast<::genium::optional<::smoke::Dates::DateStruct>*>( handle );
 }
-int64_t smoke_Dates_DateStruct_dateField_get(_baseRef handle) {
+double smoke_Dates_DateStruct_dateField_get(_baseRef handle) {
     auto struct_pointer = get_pointer<::smoke::Dates::DateStruct>(handle);
 return Conversion<std::chrono::system_clock::time_point>::toBaseRef(struct_pointer->date_field);
 }
-int64_t smoke_Dates_dateMethod(_baseRef _instance, int64_t input) {
+double smoke_Dates_dateMethod(_baseRef _instance, double input) {
     return Conversion<std::chrono::system_clock::time_point>::toBaseRef(get_pointer<std::shared_ptr<::smoke::Dates>>(_instance)->get()->date_method(Conversion<std::chrono::system_clock::time_point>::toCpp(input)))
 ;
 }
-int64_t smoke_Dates_dateAttribute_get(_baseRef _instance) {
+double smoke_Dates_dateAttribute_get(_baseRef _instance) {
     return Conversion<std::chrono::system_clock::time_point>::toBaseRef(get_pointer<std::shared_ptr<::smoke::Dates>>(_instance)->get()->get_date_attribute())
 ;
 }
-void smoke_Dates_dateAttribute_set(_baseRef _instance, int64_t newValue) {
+void smoke_Dates_dateAttribute_set(_baseRef _instance, double newValue) {
     return get_pointer<std::shared_ptr<::smoke::Dates>>(_instance)->get()->set_date_attribute(Conversion<std::chrono::system_clock::time_point>::toCpp(newValue))
 ;
 }
@@ -72,7 +72,7 @@ _baseRef smoke_Dates_DateMap_iterator(_baseRef handle) {
 void smoke_Dates_DateMap_iterator_release_handle(_baseRef iterator_handle) {
     delete reinterpret_cast<std::unordered_map<std::string, std::chrono::system_clock::time_point>::iterator*>( iterator_handle );
 }
-void smoke_Dates_DateMap_put(_baseRef handle, _baseRef key, int64_t value) {
+void smoke_Dates_DateMap_put(_baseRef handle, _baseRef key, double value) {
     (*get_pointer<std::unordered_map<std::string, std::chrono::system_clock::time_point>>(handle)).emplace(std::move(Conversion<std::string>::toCpp(key)), std::move(Conversion<std::chrono::system_clock::time_point>::toCpp(value)));
 }
 bool smoke_Dates_DateMap_iterator_is_valid(_baseRef handle, _baseRef iterator_handle) {
@@ -85,7 +85,7 @@ _baseRef smoke_Dates_DateMap_iterator_key(_baseRef iterator_handle) {
     auto& key = (*reinterpret_cast<std::unordered_map<std::string, std::chrono::system_clock::time_point>::iterator*>( iterator_handle ))->first;
     return Conversion<std::string>::toBaseRef(key);
 }
-int64_t smoke_Dates_DateMap_iterator_value(_baseRef iterator_handle) {
+double smoke_Dates_DateMap_iterator_value(_baseRef iterator_handle) {
     auto& value = (*reinterpret_cast<std::unordered_map<std::string, std::chrono::system_clock::time_point>::iterator*>( iterator_handle ))->second;
     return Conversion<std::chrono::system_clock::time_point>::toBaseRef(value);
 }
